requirements.txt
kafka-python
aiokafka
python-dotenv
pydantic
langgraph
langchain
langchain-openai
langchain-core
faker
fastapi
uvicorn
requests

# Presidio for PII detection (rule-based + ML)
presidio-analyzer
presidio-anonymizer
spacy
transformers
torch